// src/enums/ModuleTypes.ts
export enum ModuleTypes {
    TopCustomer = "top_customer",
    TextBox = "textbox",
    RecentPayments = "recent_payments",
    FeaturedPackage = "featured_package",
    GiftcardBalance = "giftcard_balance",
    ServerStatus = "server_status",
    PaymentGoal = "payment_goal",
    CommunityGoal = "community_goal",
}